refactor: 代码质量全面升级 v2.1.0
P0 严重修复:
- RobotModel: rand()→mt19937 线程安全随机数
- PathPlanner: sampleUniform 尊重关节限位替代硬编码[-π,π]
- PathPlannerOptimized: int16_t→int32_t 消除节点数溢出
- Types: PlanningResult 添加 errorMessage 字段

P1 功能完善:
- PalletizingPlanner: generateKeyConfigs 8段码垛路径生成
- PathPlanner: pruneTree 实现基于代价阈值的叶节点剪枝
- PalletizingPlanner: planTaskSequence 修复路径连续性
- PalletizingPlanner: 切换到 OptimizedPathPlanner 高性能规划器
- CollisionChecker: isPathCollisionFree 单次锁优化

P2 代码质量:
- TaskSequencer: 消除 -Wnarrowing 编译警告
- Types: 未实现枚举标注 [未实现-预留]
- 所有测试输出统一到 data/ 目录
- 删除根目录散落数据文件、备份文件、空目录

P3 工程规范:
- CMake 3.0.2→3.14, 项目名统一 HRC-PalletizingPlanner
- 13个hpp @author 修正为公司名
- CI mock模式支持 (HRC闭源库缺失时不阻断)
- .gitignore 添加 data/sim3d/*.png 例外
- examples 加入 CMake 编译目标
- README 更新: 目录结构/核心模块表/CMake版本/版本历史

编译: 7目标100%通过, 零错误零警告
测试: testPalletizingPlanner 7子测试通过
文档: PROJECT_ANALYSIS.md + ITERATION_PLAN.md

+++ include/PalletizingPlanner/PathPlannerOptimized.hpp
@@ -33,6 +33,7 @@
 #include <unordered_set>
 #include <unordered_map>
 #include <chrono>
+#include <iostream>
 #include <algorithm>
 #include <thread>
 #include <atomic>
@@ -46,13 +47,13 @@ namespace palletizing {
 struct alignas(64) OptimizedRRTNode {  // 缓存行对齐
     JointConfig config;              // 关节配置 (48 bytes)
     float costFromStart = 0.0f;      // 使用float节省内存
-    int16_t parentId = -1;           // 父节点ID
-    int16_t id;                      // 节点ID
+    int32_t parentId = -1;           // 父节点ID (int32_t避免>32767溢出)
+    int32_t id;                      // 节点ID
     uint8_t validEdge : 1;           // 边是否已验证
     uint8_t reserved : 7;            // 保留位
     
     OptimizedRRTNode() : id(-1), validEdge(0), reserved(0) {}
-    OptimizedRRTNode(int16_t nodeId, const JointConfig& cfg) 
+    OptimizedRRTNode(int32_t nodeId, const JointConfig& cfg) 
         : config(cfg), id(nodeId), validEdge(0), reserved(0) {}
 };
 
@@ -60,8 +61,8 @@ struct alignas(64) OptimizedRRTNode {  // 缓存行对齐
  * @brief 懒惰碰撞检测边
  */
 struct LazyEdge {
-    int16_t fromNode;
-    int16_t toNode;
+    int32_t fromNode;
+    int32_t toNode;
     float cost;
     bool validated = false;
     bool valid = false;
